\name{clustcorr-package}
\alias{clustcorr-package}
\alias{clustcorr}
\docType{package}
\title{
\packageTitle{clustcorr}
}
\description{
\packageDescription{clustcorr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clustcorr}
\packageIndices{clustcorr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{clustcorr}

Maintainer: \packageMaintainer{clustcorr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
print(2)
}
